+++ ytmapi-rs/src/parse.rs
@@ -322,7 +322,7 @@ fn parse_item_text(
 
 #[cfg(test)]
 mod tests {
-    use serde_json::json;
+    
 
     use crate::{process::JsonCloner, query::SearchQuery};