+++ ytmapi-rs/CHANGELOG.md
@@ -4,5 +4,6 @@
 - Added usage examples.
 - Documented public API. 
 - Removed nightly requirement.
+- Implemented all Search cases.
 # v0.0.1
 - Initial release to github / crates.io.